+++ apps/admin/src/components/dashboard/settings/settings-header.tsx
@@ -1,16 +1,13 @@
-'use client';
-
-// Node Modules
-import React, { memo } from 'react';
-import { Settings, Shield, Save } from '@repo/ui/lib/icons';
+import { memo } from 'react';
+import { Settings } from '@repo/ui/lib/icons';
 
 function SettingsHeader() {
   return (
     <div className="relative overflow-hidden rounded-xl border border-[var(--baladi-border)] bg-gradient-to-br from-[var(--baladi-primary)] via-[var(--baladi-primary)] to-[var(--baladi-secondary)] p-6 shadow-lg">
       <div className="absolute inset-0 opacity-10">
-        <div className="absolute -right-4 -top-4 h-32 w-32 rounded-full bg-white/20"></div>
+        <div className="absolute -top-4 -right-4 h-32 w-32 rounded-full bg-white/20"></div>
         <div className="absolute -bottom-8 -left-8 h-40 w-40 rounded-full bg-white/10"></div>
-        <div className="absolute right-1/4 top-1/2 h-24 w-24 rounded-full bg-white/5"></div>
+        <div className="absolute top-1/2 right-1/4 h-24 w-24 rounded-full bg-white/5"></div>
       </div>
 
       <div className="relative flex flex-col gap-6 lg:flex-row lg:items-center lg:justify-between">
