+++ apps/admin/src/hooks/useSettings/types.ts
@@ -2,27 +2,12 @@ import { ApiData } from '@/utils/types.util';
 import { Admin } from '@repo/types/admin';
 import { Config } from '@repo/types/config';
 
-export type SiteConfigBody = {
-  showPalette: boolean;
-};
-
-export type UpdateAdminPasswordSchema = {
-  body: {
+export type UpdateAdminPasswordRequest = ApiData<
+  {
     currentPassword: string;
     newPassword: string;
     confirmNewPassword: string;
-  };
-};
-
-export type GetAdminProfileRequest = ApiData<
-  undefined,
-  {
-    admin: Admin;
-  }
->;
-
-export type UpdateAdminPasswordRequest = ApiData<
-  UpdateAdminPasswordSchema['body'],
+  },
   {
     admin: Admin;
   }
@@ -35,36 +20,26 @@ export type GetAllAdminsRequest = ApiData<
   }
 >;
 
-export type CreateAdminSchema = {
-  body: {
+export type CreateAdminRequest = ApiData<
+  {
     name: string;
     email: string;
-  };
-};
-
-export type CreateAdminResponse = ApiData<
-  CreateAdminSchema['body'],
+  },
   {
     admin: Admin;
   }
 >;
 
-export type SiteConfigRequest = ApiData<
+export type GetSiteConfigRequest = ApiData<
   undefined,
   {
     config: Config;
   }
 >;
 
-export type UpdateSiteConfigSchema = {
-  body: {
-    config: Config;
-  };
-};
-
 export type UpdateSiteConfigRequest = ApiData<
-  SiteConfigBody,
   {
-    config: Config;
-  }
+    showPalette: boolean;
+  },
+  undefined
 >;
