+++ apps/api/src/models/interfaces/config.model.d.ts
@@ -1,3 +1,4 @@
-import { Config } from '@repo/types/config';
 import { Document } from 'mongoose';
+import { Config } from '@repo/types/config';
+
 export interface IConfig extends Config, Document {}
