+++ apps/api/src/routes/user.route.ts
@@ -23,8 +23,6 @@ import {
   topUsersSchema,
   updateAdminPasswordSchema,
 } from '@/validators/user.validator';
-import { createAdminSchema } from '@/validators/auth.validator';
-import { createAdmin } from '@/controllers/auth.controller';
 
 const router: Router = express.Router();
 
@@ -45,7 +43,6 @@ router.get(
 router.get('/stats', isAdmin, validate(getUserStatsSchema), getUserStats);
 router.get('/top', isAdmin, validate(topUsersSchema), getTopUsers);
 
-// Admin routes
 router.get('/admin/all', isAdmin, getAllAdmins);
 router.put(
   '/admin/password',
