+++ apps/api/src/validators/product.validator.ts
@@ -143,18 +143,6 @@ export const topProductsSchema = z.object({
     }),
 });
 
-export const updateConfigSchema = z.object({
-  body: z.object({
-    showPalette: z.boolean(),
-  }),
-});
-
-export const createConfigSchema = z.object({
-  body: z.object({
-    showPalette: z.boolean(),
-  }),
-});
-
 export const productStatsSchema = z.object({});
 
 export type GetAllProductsSchema = z.infer<typeof getAllProductsSchema>;
@@ -167,6 +155,4 @@ export type DeleteProductSchema = z.infer<typeof deleteProductSchema>;
 export type LowStockProductsSchema = z.infer<typeof lowStockProductsSchema>;
 export type TopProductsSchema = z.infer<typeof topProductsSchema>;
 export type ProductStatsSchema = z.infer<typeof productStatsSchema>;
-export type UpdateConfigSchema = z.infer<typeof updateConfigSchema>;
-export type CreateConfigSchema = z.infer<typeof createConfigSchema>;
 /******************* END: Admin Validators *******************/
