+++ packages/types/src/config.ts
@@ -1,6 +1,8 @@
 export interface Config {
   _id: string;
+
   showPalette: boolean;
+
   createdAt: Date;
   updatedAt: Date;
 }
